fix breakage of x86_64 sigaction from recent changes for mips

+++ src/internal/ksigaction.h
@@ -2,5 +2,5 @@ struct k_sigaction {
 	void (*handler)(int);
 	unsigned long flags;
 	void (*restorer)(void);
-	unsigned long mask[2];
+	unsigned mask[2];
 };
